+++ src/components/inputs/textInput/index.tsx
@@ -39,7 +39,7 @@ function TextInput({
         value={value}
         onChange={(e: ChangeEvent<HTMLInputElement>) => handleInput(e)}
         type={inputType}
-        className="w-full p-3 rounded-lg border  bg-gray-100/15 focus:ring-[#17594F] focus:border-[#17594F]"
+        className="w-full p-3 focus:outline-none rounded-lg border  bg-gray-100/15 focus:ring-[#17594F] focus:border-[#17594F]"
       />
     </div>
   );